+++ Solution3/Alignments/untitled/src/NeedlemanWunsch.java
@@ -1,6 +1,6 @@
 public abstract class NeedlemanWunsch extends AlignmentAlgorithm {
 
-    protected int[][] dp;
+    protected double[][] dp;
     protected final int gap; // linear gap
 
     public NeedlemanWunsch(Sequence seq1, Sequence seq2, Matrix scoringMatrix, GapPenalty gapPenalty) {
@@ -21,17 +21,17 @@ public abstract class NeedlemanWunsch extends AlignmentAlgorithm {
     }
 
     // i/j sind dp-Indizes (1..n / 1..m)
-    protected abstract int computeCellValue(int i, int j);
+    protected abstract double computeCellValue(int i, int j);
 
-    protected int diag(int i, int j) {
+    protected double diag(int i, int j) {
         return dp[i - 1][j - 1] + score(i - 1, j - 1);
     }
 
-    protected int up(int i, int j) {
+    protected double up(int i, int j) {
         return dp[i - 1][j] + gap;
     }
 
-    protected int left(int i, int j) {
+    protected double left(int i, int j) {
         return dp[i][j - 1] + gap;
     }
 }
