+++ Solution3/Alignments/untitled/src/NeedlemanWunschGlobal.java
@@ -9,7 +9,7 @@ public class NeedlemanWunschGlobal extends NeedlemanWunsch {
         int n = seq1.getSequence().length();
         int m = seq2.getSequence().length();
 
-        dp = new int[n + 1][m + 1];
+        dp = new double[n + 1][m + 1];
 
         dp[0][0] = 0;
         for (int i = 1; i <= n; i++) dp[i][0] = dp[i - 1][0] + gap;
@@ -17,10 +17,10 @@ public class NeedlemanWunschGlobal extends NeedlemanWunsch {
     }
 
     @Override
-    protected int computeCellValue(int i, int j) {
-        int diag = diag(i, j);
-        int up = up(i, j);
-        int left = left(i, j);
+    protected double computeCellValue(int i, int j) {
+        double diag = diag(i, j);
+        double up = up(i, j);
+        double left = left(i, j);
         return Math.max(diag, Math.max(up, left));
     }
 
@@ -51,7 +51,7 @@ public class NeedlemanWunschGlobal extends NeedlemanWunsch {
 
         String aligned1 = a1.reverse().toString();
         String aligned2 = a2.reverse().toString();
-        int finalScore = dp[seq1.getSequence().length()][seq2.getSequence().length()];
+        double finalScore = dp[seq1.getSequence().length()][seq2.getSequence().length()];
 
         int n = seq1.getSequence().length();
         int m = seq2.getSequence().length();
